refactor UI auth code
